+++ ambassador/entrypoint.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 export LC_ALL=C.UTF-8
 export LANG=C.UTF-8
@@ -11,68 +11,60 @@ fi
 
 APPDIR=${APPDIR:-/application}
 
-VERSION=$(python3 -c 'from ambassador.VERSION import Version; print(Version)')
-
-pids=()
-
-log () {
-    now=$(date +"%Y-%m-%d %H:%M:%S")
-    echo "$now AMBASSADOR: $@"
-}
+pids=""
 
 diediedie() {
     NAME=$1
     STATUS=$2
 
     if [ $STATUS -eq 0 ]; then
-        log "$NAME claimed success, but exited \?\?\?\?"
+        echo "AMBASSADOR: $NAME claimed success, but exited \?\?\?\?"
     else
-        log "$NAME exited with status $STATUS"
+        echo "AMBASSADOR: $NAME exited with status $STATUS"
     fi
 
-    log "Here's the envoy.json we were trying to run with:"
+    echo "Here's the envoy.json we were trying to run with:"
     LATEST="$(ls -v /etc/envoy*.json | tail -1)"
     if [ -e "$LATEST" ]; then
-        cat $LATEST
+        cat "$LATEST"
     else
-        log "No config generated."
+        echo "No config generated."
     fi
 
-    log "shutting down"
+    echo "AMBASSADOR: shutting down"
     exit 1
 }
 
 handle_chld() {
-    local tmp=()
-
-    for (( i=0; i<${#pids[@]}; ++i )); do
-        split=(${pids[$i]//;/ })    # the space after the trailing / is critical!
-        pid=${split[0]}
-        name=${split[1]}
-
-        if [ ! -d /proc/$pid ]; then
-            wait $pid
+    trap - CHLD
+    local tmp
+    for entry in $pids; do
+        local pid="${entry%:*}"
+        local name="${entry#*:}"
+        if [ ! -d "/proc/${pid}" ]; then
+            wait "${pid}"
             STATUS=$?
-            diediedie "$name" "$STATUS"
+            # echo "AMBASSADOR: $name exited: $STATUS"
+            # echo "AMBASSADOR: shutting down"
+            diediedie "${name}" "$STATUS"
         else
-            tmp+=(${pids[i]})
+            tmp="${tmp:+${tmp} }${entry}"
         fi
     done
 
-    pids=(${tmp[@]})
+    pids="$tmp"
+    trap "handle_chld" CHLD
 }
 
 handle_int() {
-    log "Exiting due to Control-C"
+    echo "Exiting due to Control-C"
 }
 
 set -o monitor
 trap "handle_chld" CHLD
 trap "handle_int" INT
 
-log "starting initial sync"
-
-/usr/bin/python3 "$APPDIR/kubewatch.py" sync "$CONFIG_DIR" /etc/envoy.json 
+/usr/bin/python3 "$APPDIR/kubewatch.py" sync "$CONFIG_DIR" /etc/envoy.json
 
 STATUS=$?
 
@@ -80,26 +72,18 @@ if [ $STATUS -ne 0 ]; then
     diediedie "kubewatch sync" "$STATUS"
 fi
 
-if [ -z "$AMBASSADOR_NO_DIAGD" ]; then
-    log "starting diagd"
-    diagd --no-debugging "$CONFIG_DIR" &
-    pids+=("$!;diagd")
-fi
+echo "AMBASSADOR: starting diagd"
+diagd --no-debugging "$CONFIG_DIR" &
+pids="${pids:+${pids} }$!:diagd"
 
-if [ -z "$AMBASSADOR_NO_ENVOY" ]; then
-    log "starting Envoy"
-    /usr/bin/python3 "$APPDIR/hot-restarter.py" "$APPDIR/start-envoy.sh" &
-    RESTARTER_PID="$!"
-    pids+=("${RESTARTER_PID};envoy")
+echo "AMBASSADOR: starting Envoy"
+/usr/bin/python3 "$APPDIR/hot-restarter.py" "$APPDIR/start-envoy.sh" &
+RESTARTER_PID="$!"
+pids="${pids:+${pids} }${RESTARTER_PID}:envoy"
 
-    log "restarter PID $RESTARTER_PID"
-    log "starting kubewatch"
+/usr/bin/python3 "$APPDIR/kubewatch.py" watch "$CONFIG_DIR" /etc/envoy.json -p "${RESTARTER_PID}" &
+pids="${pids:+${pids} }$!:kubewatch"
 
-    if [ -z "$AMBASSADOR_NO_KUBEWATCH" ]; then
-        /usr/bin/python3 "$APPDIR/kubewatch.py" watch "$CONFIG_DIR" /etc/envoy.json -p "${RESTARTER_PID}" &
-        pids+=("$!;kubewatch")
-    fi
-fi
-
-log "ready for action"
+echo "AMBASSADOR: waiting"
+echo "PIDS: $pids"
 wait