+++ end-to-end/003-headers-and-host/test.sh
@@ -11,14 +11,16 @@ PATH="${ROOT}:${PATH}"
 
 source ${ROOT}/utils.sh
 
-shred_and_reclaim
+initialize_cluster
 
 kubectl cluster-info
 
 kubectl apply -f k8s/ambassador.yaml
+kubectl apply -f k8s/cors.yaml
 kubectl apply -f k8s/demo1.yaml
 kubectl apply -f k8s/demo2.yaml
 kubectl apply -f ${ROOT}/ambassador-deployment.yaml
+kubectl run demotest --image=dwflynn/demotest:0.0.1 -- /bin/sh -c "sleep 3600"
 
 set +e +o pipefail
 
@@ -26,6 +28,7 @@ wait_for_pods
 
 CLUSTER=$(cluster_ip)
 APORT=$(service_port ambassador)
+DEMOTEST_POD=$(demotest_pod)
 
 BASEURL="http://${CLUSTER}:${APORT}"
 
@@ -38,7 +41,11 @@ if ! check_diag "$BASEURL" 1 "No canary active"; then
     exit 1
 fi
 
-if ! demotest.py "$BASEURL" demo-1.yaml; then
+if ! kubectl exec -i $DEMOTEST_POD -- python3 demotest.py "$BASEURL" /dev/fd/0 < cors.yaml; then
+    exit 1
+fi
+
+if ! kubectl exec -i $DEMOTEST_POD -- python3 demotest.py "$BASEURL" /dev/fd/0 < demo-1.yaml; then
     exit 1
 fi
 
@@ -51,7 +58,9 @@ wait_for_demo_weights "$BASEURL" x-demo-mode=canary 50 50
 #     exit 1
 # fi
 
-if ! demotest.py "$BASEURL" demo-2.yaml; then
+sleep 5
+
+if ! kubectl exec -i $DEMOTEST_POD -- python3 demotest.py "$BASEURL" /dev/fd/0 < demo-2.yaml; then
     exit 1
 fi
 
@@ -64,7 +73,9 @@ wait_for_demo_weights "$BASEURL" x-demo-mode=canary 100
 #     exit 1
 # fi
 
-if ! demotest.py "$BASEURL" demo-3.yaml; then
+sleep 5
+
+if ! kubectl exec -i "$DEMOTEST_POD" -- python3 demotest.py "$BASEURL" /dev/fd/0 < demo-3.yaml; then
     exit 1
 fi