+++ scripts/travis-build.sh
@@ -8,6 +8,8 @@ if [ $(echo "$TRAVIS_BRANCH" | egrep -c '^v[0-9][0-9\.]*$') -gt 0 ]; then
     exit 0
 fi
 
+export AWS_DEFAULT_REGION=us-east-1
+
 env | grep TRAVIS | sort
 npm version
 aws --version
@@ -45,11 +47,18 @@ onmaster () {
 }
 
 # Do we have any non-doc changes?
-echo "======== Diff summary"
-git diff --stat "$TRAVIS_COMMIT_RANGE"
+DIFF_RANGE=${TRAVIS_COMMIT_RANGE:-HEAD^}
 
-nondoc_changes=$(git diff --name-only "$TRAVIS_COMMIT_RANGE" | grep -v '^docs/' | wc -l | tr -d ' ')
-doc_changes=$(git diff --name-only "$TRAVIS_COMMIT_RANGE" | grep -e '^docs/' | wc -l | tr -d ' ')
+echo "======== Diff summary ($DIFF_RANGE)"
+git diff --stat "$DIFF_RANGE"
+
+nondoc_changes=$(git diff --name-only "$DIFF_RANGE" | grep -v '^docs/' | wc -l | tr -d ' ')
+doc_changes=$(git diff --name-only "$DIFF_RANGE" | grep -e '^docs/' | wc -l | tr -d ' ')
+
+# # Use this hack to force a doc-only build. Ew.
+# nondoc_changes=0
+# doc_changes=1
+# TRAVIS_COMMIT_RANGE=some
 
 # Default VERSION to _the current version of Ambassador._
 VERSION=$(python scripts/versioner.py)
@@ -66,11 +75,11 @@ if [ \( -z "$TRAVIS_COMMIT_RANGE" \) -o \( $nondoc_changes -gt 0 \) ]; then
         git checkout ${TRAVIS_BRANCH}
 
         # ...make sure we're interacting with our official Docker repo...
-        DOCKER_REGISTRY="datawire"
+        DOCKER_REGISTRY="quay.io/datawire"
 
         set +x
         echo "+docker login..."
-        $ECHO docker login -u "${DOCKER_USERNAME}" -p "${DOCKER_PASSWORD}"
+        $ECHO docker login -u "${DOCKER_USERNAME}" -p "${DOCKER_PASSWORD}" quay.io
         set -x
 
         # We _won't_ try to figure out a magic prebuild number for real builds.
@@ -98,18 +107,25 @@ if [ \( -z "$TRAVIS_COMMIT_RANGE" \) -o \( $nondoc_changes -gt 0 \) ]; then
         # ...and, if we're on master, tag this version...
         $ECHO make VERSION=${VERSION} tag
 
-        # ...and push the tag.
+        # ...push the tag...
         $ECHO git push --tags https://d6e-automation:${GH_TOKEN}@github.com/datawire/ambassador.git master
+
+        # ...and update our stable.txt.
+        printf "${VERSION}" > stable.txt
+
+        $ECHO aws s3api put-object \
+            --bucket datawire-static-files \
+            --key ambassador/stable.txt \
+            --body stable.txt
     else
         # If not on master, don't tag...
         echo "not on master; not tagging"
 
         # ...and push app.json to testapp.json for later examination.
-        SCOUT_KEY=testapp.json
+        SCOUT_KEY=testapp.json 
     fi
 
     # Push new info to AWS
-    export AWS_DEFAULT_REGION=us-east-1
     $ECHO aws s3api put-object \
         --bucket scout-datawire-io \
         --key ambassador/$SCOUT_KEY \
@@ -128,7 +144,7 @@ if [ $doc_changes -gt 0 ]; then
     # Yes, so we'll run a doc build, for which we always use the Datawire registry.
     # (why? 'cause there's no way to figure out WTF domain name Netlify will push to
     # at this point)
-    DOCKER_REGISTRY=datawire
+    DOCKER_REGISTRY=quay.io/datawire
 
     if onmaster; then
         # If on master, we publish instead of just leaving everything in draft mode.
@@ -148,10 +164,9 @@ if [ $doc_changes -gt 0 ]; then
 
     $ECHO make VERSION=${VERSION} travis-website
 
-    $ECHO docs/node_modules/.bin/netlify --access-token ${NETLIFY_TOKEN} \
-        deploy --path docs/_book \
-               --site-id datawire-ambassador \
-               ${NETLIFY_DRAFT}
+    $ECHO docs/node_modules/.bin/netlify --access-token "${NETLIFY_TOKEN}" \
+        deploy $NETLIFY_DRAFT --path docs/_book \
+               --site-id datawire-ambassador
 else
     echo "Not building docs for $VERSION; no doc changes"
 fi